index.js: bot now only replies to "/ambience", set volume: false
The logic to check if a user had the role permission to summon the bot was previosuly placed before the check for the exact message = "/ambience", thus the bot would spontaneously reply to all messages. Place it after the check for the right message, so it doesn't spam anymore.

Set volume: false within play() in the discord.js api, as volume cntrol isnt necessary (the clients can set this, and there's no reason to alter the volume of the input stream before the client does-- desktop audio should be at 100% of input volume)

also clarify some comments!